+++ mrp_subcontracting/i18n/es.po
@@ -7,7 +7,7 @@ msgstr ""
 "Project-Id-Version: Odoo Server 8.0\n"
 "Report-Msgid-Bugs-To: \n"
 "POT-Creation-Date: 2014-12-10 15:03+0000\n"
-"PO-Revision-Date: 2014-12-10 16:09+0100\n"
+"PO-Revision-Date: 2014-12-15 12:28+0100\n"
 "Last-Translator: Alfredo <alfredodelafuente@avanzosc.com>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -19,7 +19,7 @@ msgstr ""
 #: code:addons/mrp_subcontracting/models/mrp_subcontrating.py:114
 #, python-format
 msgid "Company warehouse not found"
-msgstr "No encontrado el almacén de la compañía"
+msgstr "No se ha encontrado el almacén de la compañía"
 
 #. module: mrp_subcontracting
 #: model:ir.actions.act_window,name:mrp_subcontracting.action_purchases_per_production
