+++ mrp_subcontracting/views/mrp_production_view.xml
@@ -1,61 +1,6 @@
 <?xml version="1.0"?>
 <openerp>
 <data>
-    <record model="ir.ui.view" id="mrp_production_workcenter_subcontracting_form_view_inherit">
-        <field name="name">mrp.production.workcenter.line.subcontracting.form</field>
-        <field name="model">mrp.production.workcenter.line</field>
-        <field name="inherit_id" ref="mrp_operations.mrp_production_workcenter_form_view_inherit"/>
-        <field name="arch" type="xml">
-            <field name="sequence" position="after">
-                <field name="external" invisible="1"/>
-            </field>
-			<page string="Information" position="after">
-				<page string="Subcontracting" attrs="{'invisible': [('external', '=', False)]}">
-					<group>
-						<field name="purchase_order"/>
-						<field name="out_picking"/>
-                        <field name="procurement_order" />
-						<field name="in_picking"/>
-					</group>
-				</page>
-			</page>
-        </field>
-    </record>
-    <record model="ir.ui.view" id="mrp_production_workcenter_tree_view_inh_subcontracting">
-        <field name="name">mrp.production.workcenter.tree.view.inh.subcontracting</field>
-        <field name="model">mrp.production.workcenter.line</field>
-        <field name="inherit_id" ref="mrp_operations.mrp_production_workcenter_tree_view_inherit"/>
-        <field name="arch" type="xml">
-            <field name="production_id" position="after">
-                <field name="external" />
-            </field>
-        </field>
-    </record>
-
-        <record id="action_purchases_per_production" model="ir.actions.act_window">
-            <field name="name">Created Purchases</field>
-            <field name="res_model">purchase.order</field>
-            <field name="view_mode">tree,form</field>
-            <field name="domain">[('mrp_production','=',active_id)]
-            </field>
-        </record>
-
-        <record id="action_outpickings_per_production" model="ir.actions.act_window">
-            <field name="name">Out Pickings</field>
-            <field name="res_model">stock.picking</field>
-            <field name="view_mode">tree,form</field>
-            <field name="domain">[('mrp_production','=',active_id),('picking_type_id.code','=','outgoing')]
-            </field>
-        </record>
-
-        <record id="action_inpickings_per_production" model="ir.actions.act_window">
-            <field name="name">In Pickings</field>
-            <field name="res_model">stock.picking</field>
-            <field name="view_mode">tree,form</field>
-            <field name="domain">[('mrp_production','=',active_id),('picking_type_id.code','=','incoming')]
-            </field>
-        </record>
-
     <record id="mrp_production_form_view_inh_subcontracting" model="ir.ui.view">
         <field name="name">mrp.production.form.view.inh.subcontracting</field>
         <field name="model">mrp.production</field>
@@ -92,7 +37,6 @@
             </div>
         </field>
     </record>
-
     <record id="mrp_production_form_view2_inh_subcontracting" model="ir.ui.view">
         <field name="name">mrp.production.form.view2.inh.subcontracting</field>
         <field name="model">mrp.production</field>
@@ -110,7 +54,5 @@
             </page>
         </field>
     </record>
-        
-
 </data>
 </openerp>
