+++ mrp_subcontracting/views/mrp_routing_workcenter_view.xml
@@ -6,24 +6,22 @@
             <field name="model">mrp.routing.workcenter</field>
             <field name="inherit_id" ref="mrp.mrp_routing_workcenter_tree_view"/>
             <field name="arch" type="xml">
-            	<field name="workcenter_id" position="after">
-            		<field name="external"/>
-            	</field>
+                <field name="workcenter_id" position="after">
+                    <field name="external"/>
+                </field>
             </field>
         </record>
-
         <record id="mrp_routing_workcenter_inherit_form_view" model="ir.ui.view">
             <field name="name">mrp.routing.workcenter.inherit.form</field>
             <field name="model">mrp.routing.workcenter</field>
             <field name="inherit_id" ref="mrp.mrp_routing_workcenter_form_view"/>
             <field name="arch" type="xml">
-            	<field name="workcenter_id" position="after">
-            		<field name="external"/>
-            		<field name="semifinished_id" attrs="{'invisible': [('external', '=', False)]}"/>
-            		<field name="picking_type_id" attrs="{'invisible': [('external', '=', False)]}"/>
-            	</field>
+                <field name="workcenter_id" position="after">
+                    <field name="external"/>
+                    <field name="semifinished_id" attrs="{'invisible': [('external', '=', False)]}"/>
+                    <field name="picking_type_id" attrs="{'invisible': [('external', '=', False)]}"/>
+                </field>
             </field>
         </record>
-
     </data>
 </openerp>
