+++ mrp_subcontracting/views/purchase_order_view.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<openerp>
+<data>
+        <record id="action_purchases_per_production" model="ir.actions.act_window">
+            <field name="name">Created Purchases</field>
+            <field name="res_model">purchase.order</field>
+            <field name="view_mode">tree,form</field>
+            <field name="domain">[('mrp_production','=',active_id)]
+            </field>
+        </record>
+</data>
+</openerp>
